Rebuild chains to - mostly - chain from /blog/*/
This is a large commit due to my inadequate discipline when it comes
to remembering to commit.

The majority of it is based in re-chaining Posts.pm so that its
actions chain from Blog.pm's action(s).

There is also a utility action for collecting a single post, and I am
migrating it to FormFu, although the YAML files for that have not been
made yet.