Fix posts/add
This will kick you off if this is not your blog. Otherwise it uses
FormFu to show the same form as before.

NOTE: this needs integrating with a decent warning system.

Other note: The forms have no CSS at the moment. One big TODO is to
create form field templates and use those to override the defaults
that FormFu uses to render a form. This will make styling the form
much easier.