Rename Blog.pm to Blogs.pm
Why? Well, there is the 'blog' action in root, which captures
/blog/* and forwards to /blog/load_blog. Unfortunately, this
simply forwards to itself and causes deep recursion.

In order to separate the namespaces I am basically doing the
same thing we've already done with posts, which is to have the
/blog/ namespace for a chain starting in root, and the /blogs/
namespace for various actions not chained from there.