Turns out we can use an indicator in the form YAML
The default behaviour, when we click 'preview', is that
the *page* is submitted, but FormFu says the form itself
is not. Thus, if we click preview, the form is submitted
such that the template reloads and shows the preview and
the form keeps its data, but FormFu does not pass its
'submitted_and_valid' test.